+++ core/sagas/modals.js
@@ -1,24 +1,12 @@
 import { call, put, takeEvery } from "redux-saga/effects";
-import { fetchBlueprintContentsApi, fetchModalCreateImageTypesApi, fetchSourceInfoApi } from "../apiCalls";
+import { fetchModalCreateImageTypesApi, fetchSourceInfoApi } from "../apiCalls";
 
 import {
-  FETCHING_MODAL_EXPORT_BLUEPRINT_CONTENTS,
-  setModalExportBlueprintContents,
   fetchingModalCreateImageTypesSuccess,
   FETCHING_MODAL_MANAGE_SOURCES_CONTENTS,
   setModalManageSourcesContents,
   modalManageSourcesFailure
 } from "../actions/modals";
-
-function* fetchModalBlueprintContents(action) {
-  try {
-    const { blueprintName } = action.payload;
-    const response = yield call(fetchBlueprintContentsApi, blueprintName);
-    yield put(setModalExportBlueprintContents(response.components));
-  } catch (error) {
-    console.log("Error in loadModalBlueprintSaga");
-  }
-}
 
 function* fetchModalManageSourcesContents() {
   try {
@@ -40,7 +28,6 @@ function* fetchModalCreateImageTypes() {
 }
 
 export default function*() {
-  yield takeEvery(FETCHING_MODAL_EXPORT_BLUEPRINT_CONTENTS, fetchModalBlueprintContents);
   yield takeEvery(FETCHING_MODAL_MANAGE_SOURCES_CONTENTS, fetchModalManageSourcesContents);
   yield* fetchModalCreateImageTypes();
 }
